@@ -1,15 +1,16 @@
-import 'package:flutter/material.dart';
-import 'package:get/get.dart';
 import 'package:banhang/route/app_page.dart';
 import 'package:banhang/route/app_route.dart';
 import 'package:banhang/theme/app_theme.dart';
+import 'package:flutter/material.dart';
+import 'package:get/get.dart';
 
 void main() {
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
+
   @override
   Widget build(BuildContext context) {
     return GetMaterialApp(
@@ -18,7 +19,6 @@ class MyApp extends StatelessWidget {
       debugShowCheckedModeBanner: false,
       theme: AppTheme.lightTheme,
       themeMode: ThemeMode.light,
-
     );
   }
 }
